4rd commit

+++ main.go
@@ -6,17 +6,17 @@ import (
 	"log"
 	"os"
 	"path/filepath"
-	"time"
+
 	"github.com/alexpfx/go_shell/internal/goshell"
 	"github.com/urfave/cli/v2"
 )
 
-const defaultBackupFile = "bkp"
+const defaultBackupFile = "password-store.plain"
 
 func main() {
 	homeDir, _ := os.UserHomeDir()
 	var defaultPasswordStore = filepath.Join(homeDir, ".password-store/")
-
+	var debugMode bool
 	app := &cli.App{
 		Name:  "go_shell",
 		Usage: "scripts de linux",
@@ -25,6 +25,36 @@ func main() {
 				Name:  "pass",
 				Usage: "comandos relativos ao comando pass",
 				Subcommands: []*cli.Command{
+					{
+						Name: "restore",
+						Action: func(c *cli.Context) error {
+							target := c.String("backup_file")
+							dFile := goshell.DecryptFile(target + ".gpg")
+							if (debugMode){
+								fmt.Println(dFile)
+							}
+
+							return nil
+						},
+						Flags: []cli.Flag{
+							&cli.StringFlag{
+								Name:    "backup_file",
+								Aliases: []string{"t"},
+								Value:   defaultBackupFile,
+							},
+							&cli.StringFlag{
+								Name:    "password-store",
+								Aliases: []string{"d"},
+								EnvVars: []string{"PASSWORD_STORE_DIR"},
+								Value:   defaultPasswordStore,
+							},
+							&cli.StringFlag{
+								Name: "public-key",
+								Aliases: []string{"k"},
+								Value: "",
+							},
+						},
+					},
 					{
 						Name: "backup",
 						Flags: []cli.Flag{
@@ -35,15 +65,15 @@ func main() {
 								Value:   defaultPasswordStore,
 							},
 							&cli.StringFlag{
-								Name:    "target",
+								Name:    "backup_file",
 								Aliases: []string{"t"},
-								Value:   fmt.Sprintf("%s_%d", defaultBackupFile, time.Now().Local().Unix()),
+								Value:   defaultBackupFile,
 							},
 						},
 						Action: func(c *cli.Context) error {
 
 							passwordStore := c.String("password-store")
-							target := c.String("target")
+							target := c.String("backup_file")
 
 							list := goshell.GetGpgFilePaths(passwordStore)
 
@@ -53,26 +83,36 @@ func main() {
 								password := goshell.OpenPass(passwordStore, pn)
 								allPassInfos = append(allPassInfos, password)
 							}
-							
-							
-							if goshell.CheckFileExists(target){
-								log.Fatal("File exists: ", target)
+
+							if goshell.CheckFileExists(target) {
+								if target != defaultPasswordStore {
+									log.Fatal("File exists: ", target)
+								}
 							}
 
 							ioutil.WriteFile(target, goshell.ToJsonStr(allPassInfos), 0644)
 							_, err := goshell.EncryptFile(target, "")
-							if err != nil {								
+							if err != nil {
 								log.Fatal(err)
 							}
-							err = os.Remove(target)
-							if err != nil{
+							if !debugMode {
+								err = os.Remove(target)
+							}
+							if err != nil {
 								log.Fatal(err)
 							}
-														
+
 							return nil
 						},
 					},
 				},
+			},
+		},
+		Flags: []cli.Flag{
+			&cli.BoolFlag{
+				Name:        "debug",
+				Aliases:     []string{"D"},
+				Destination: &debugMode,
 			},
 		},
 		Action: func(c *cli.Context) error {
